fix(test): mock @native-html/render for Jest ESM compatibility
PaintedUsernameHtml imports @native-html/render, which pulls in ESM-only
dependencies that Jest cannot parse. Route the module to a shared mock so
transitive test imports no longer fail the full suite and chat perf job.
