+++ jest.config.js
@@ -14,6 +14,7 @@ const jestConfig = {
   moduleNameMapper: {
     '^@app/(.*)$': '<rootDir>/src/$1',
     '^@modules/(.*)$': '<rootDir>/modules/$1',
+    '^@native-html/render$': '<rootDir>/__mocks__/@native-html/render.tsx',
     '\\.otf$': '<rootDir>/__mocks__/fileMock.js',
     '^@bacons/apple-colors$': '<rootDir>/__mocks__/@bacons/apple-colors.ts',
     '^react-native-legal$': '<rootDir>/__mocks__/react-native-legal.ts',
